mcp-agent-control-plane/pyproject.toml
[project]
name = "mcp-agent-control-plane"
version = "0.1.0"
description = "mcp-agent-control-plane MCP server built with FastMCP"
authors = [
    {name = "Rogelio Rojas", email = "rogelioernesto_3@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "kubernetes>=35.0.0",
    "psycopg[binary,pool]>=3.3.2",
    "jinja2>=3.1.6",
    "kubernetes-asyncio>=35.0.0",
    "langchain-openai>=1.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
"mcp-agent-control-plane-server" = "src.main:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
